+++ src/Oro/Bundle/DataGridBundle/Datagrid/RequestParameterBagFactory.php
@@ -2,8 +2,12 @@
 
 namespace Oro\Bundle\DataGridBundle\Datagrid;
 
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RequestStack;
 
+/**
+ * Creates datagrid ParameterBag based on the request query.
+ */
 class RequestParameterBagFactory
 {
     const DEFAULT_ROOT_PARAM = 'grid';
@@ -33,16 +37,28 @@ class RequestParameterBagFactory
      *
      * @return array
      */
-    public function fetchParameters($gridParameterName = self::DEFAULT_ROOT_PARAM)
+    public function fetchParameters(string $gridParameterName = self::DEFAULT_ROOT_PARAM): array
     {
         $request = $this->requestStack->getCurrentRequest();
-        $parameters = $request ? $request->get($gridParameterName, []) : [];
 
+        return $request ? $this->fetchParametersFromRequest($request, $gridParameterName) : [];
+    }
+
+    /**
+     * @param Request $request
+     * @param string $gridParameterName
+     * @return array
+     */
+    public function fetchParametersFromRequest(
+        Request $request,
+        string $gridParameterName = self::DEFAULT_ROOT_PARAM
+    ): array {
+        $parameters = $request->get($gridParameterName, []);
         if (!is_array($parameters)) {
             $parameters = [];
         }
 
-        $minifiedParameters = $this->getMinifiedParameters($gridParameterName);
+        $minifiedParameters = $this->getMinifiedParameters($request, $gridParameterName);
         if ($minifiedParameters) {
             $parameters[ParameterBag::MINIFIED_PARAMETERS] = $minifiedParameters;
         }
@@ -54,7 +70,7 @@ class RequestParameterBagFactory
      * @param string $gridParameterName
      * @return ParameterBag
      */
-    public function createParameters($gridParameterName = self::DEFAULT_ROOT_PARAM)
+    public function createParameters(string $gridParameterName = self::DEFAULT_ROOT_PARAM): ParameterBag
     {
         $parameters = $this->fetchParameters($gridParameterName);
 
@@ -62,19 +78,32 @@ class RequestParameterBagFactory
     }
 
     /**
+     * @param Request $request
      * @param string $gridParameterName
-     * @return null
+     * @return ParameterBag
      */
-    protected function getMinifiedParameters($gridParameterName)
+    public function createParametersFromRequest(
+        Request $request,
+        string $gridParameterName = self::DEFAULT_ROOT_PARAM
+    ): ParameterBag {
+        $parameters = $this->fetchParametersFromRequest($request, $gridParameterName);
+
+        return new $this->parametersClass($parameters);
+    }
+
+    /**
+     * @param Request $request
+     * @param string $gridParameterName
+     * @return array
+     */
+    private function getMinifiedParameters(Request $request, string $gridParameterName): array
     {
-        $request = $this->requestStack->getCurrentRequest();
-        $gridData = $request ? $request->get(self::DEFAULT_ROOT_PARAM, []) : [];
-        if (empty($gridData[$gridParameterName])) {
-            return null;
+        $gridData = $request->get(self::DEFAULT_ROOT_PARAM, []);
+
+        if (!empty($gridData[$gridParameterName]) && is_string($gridData[$gridParameterName])) {
+            parse_str($gridData[$gridParameterName], $parameters);
         }
 
-        parse_str($gridData[$gridParameterName], $parameters);
-
-        return $parameters;
+        return $parameters ?? [];
     }
 }