+++ src/Oro/Bundle/DataGridBundle/Exception/DatagridNotFoundException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Oro\Bundle\DataGridBundle\Exception;
+
+/**
+ * Thrown when datagrid is not found.
+ */
+class DatagridNotFoundException extends \LogicException implements DatagridException
+{
+}