+++ src/Oro/Bundle/LayoutBundle/Provider/Image/ThemeImagePlaceholderProvider.php
@@ -6,6 +6,7 @@ use Liip\ImagineBundle\Imagine\Cache\CacheManager;
 use Oro\Bundle\LayoutBundle\Layout\LayoutContextHolder;
 use Oro\Component\Layout\Extension\Theme\Model\ThemeManager;
 use Oro\Component\Layout\LayoutContext;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 /**
  * Provides the path to the image placeholder for the actual theme.
@@ -45,14 +46,20 @@ class ThemeImagePlaceholderProvider implements ImagePlaceholderProviderInterface
     /**
      * {@inheritdoc}
      */
-    public function getPath(string $filter): ?string
+    public function getPath(string $filter, int $referenceType = UrlGeneratorInterface::ABSOLUTE_PATH): ?string
     {
         $imagePlaceholders = $this->getImagePlaceholders();
         if (!isset($imagePlaceholders[$this->placeholderName])) {
             return null;
         }
 
-        return $this->imagineCacheManager->getBrowserPath($imagePlaceholders[$this->placeholderName], $filter);
+        return $this->imagineCacheManager->generateUrl(
+            $imagePlaceholders[$this->placeholderName],
+            $filter,
+            [],
+            null,
+            $referenceType
+        );
     }
 
     /**